backend/requirements.txt
# sinX Threat Hunter - Backend Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# Redis & Caching
redis==5.0.1
aioredis==2.0.1

# Background Tasks
celery==5.3.6
apscheduler==3.10.4

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Log Parsing
python-dateutil==2.8.2
pytz==2024.1

# Threat Intelligence
stix2==3.0.1
taxii2-client==2.3.0

# GeoIP & Enrichment
geoip2==4.7.0
dnspython==2.5.0
python-whois==0.8.0

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
